Rearrange files, update doc
